backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-whois
python-dotenv
aiohttp
beautifulsoup4
lxml
pandas
tenacity
python-dateutil
tldextract
cachetools
